Dodanie Tabu Search i testów.

+++ Data.h
@@ -12,7 +12,7 @@ class Data
 {
 public:
 	string instance;
-	long int number;
+	int number;
 	int** graph;
 	int sumOfEdgesFromRandomPath;
